Move action helpers to another file
